stable waypoint, documentation pending, memory moce pending
