+++ datafusion/catalog/src/session.rs
@@ -132,6 +132,9 @@ pub trait Session: Send + Sync {
 
     /// Returns a mutable reference to [`TableOptions`]
     fn table_options_mut(&mut self) -> &mut TableOptions;
+
+    /// Get a new TaskContext to run in this session
+    fn task_ctx(&self) -> Arc<TaskContext>;
 }
 
 /// Create a new task context instance from Session
